BACKEND/requirements.minimal.txt
# Minimal requirements for Social Feed API
# Install with: pip install -r requirements.minimal.txt

Django==5.2.3
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0

# Optional (for production):
# gunicorn==23.0.0
# psycopg2-binary==2.9.10  # PostgreSQL
# python-decouple==3.8  # Environment variables
